src/requirements.txt
sodapy==2.1.0
pandas==1.2.4
psycopg2-binary==2.8.6
SQLAlchemy==1.4.11
